+++ fuseimg/datasets/isic.py
@@ -17,6 +17,7 @@ from fuse.data.ops.ops_aug_common import OpSample
 from fuse.data.ops.ops_read import OpReadDataframe
 from fuse.data.ops.ops_common import OpLambda, OpOverrideNaN, OpConcat
 from fuseimg.data.ops.color import OpToRange
+from fuse.data.ops.ops_aug_tabular import OpAugOneHotWithProb
 
 from fuse.utils import NDict
 
@@ -365,82 +366,5 @@ class OpEncodeMetaData(OpBase):
             age_one_hot[encode_idx] = 1
 
             sample_dict[f"{out_prefix}.age"] = age_one_hot
-
-        return sample_dict
-
-
-class OpAugReplaceWithProb(OpBase):
-    def __call__(self, sample_dict: NDict, key: str, value: Any, prob: float) -> NDict:
-
-        if prob < 0 or prob > 1:
-            raise Exception("prob should be between 0 and 1")
-
-        if random.random() < prob:  # can also use 'pyprob' library
-            sample_dict[key] = value
-
-        return sample_dict
-
-
-class OpAugOneHotWithProb(OpBase):
-    """
-    Apply an augmentation for an one-hot encoding vector with the following modes:
-
-    mode == "default":
-        Switch
-    """
-
-    def __call__(
-        self,
-        sample_dict: NDict,
-        key: str,
-        prob: float,
-        idx: Optional[int] = None,
-        freeze_indices: List[int] = [],
-        mode: str = "default",
-    ) -> NDict:
-        """
-        :param key: key for the one-hot vector
-        :param idx: idx to be change to 1
-        :param prob: the probability the the functionality will happen
-        :param mode: see class desc
-        """
-
-        if prob < 0 or prob > 1:
-            raise Exception("prob should be between 0 and 1")
-
-        supported_modes = ["default", "ranking"]
-        if mode not in supported_modes:
-            raise Exception(f"mode ({mode}) should be in supported modes ({supported_modes}).")
-
-        if mode != "default" and idx is not None:
-            raise Exception("specify idx only in default mode")
-
-        if mode == "default" and idx is None:
-            raise Exception("in 'default' mode, idx must be provided.")
-
-        if random.random() < prob:  # can also use 'pyprob' library
-            one_hot = sample_dict[key]
-
-            if mode == "default":
-                # TODO add freeze
-                one_hot = np.zeros_like(one_hot)
-                one_hot[idx] = 1
-                sample_dict[key] = one_hot
-
-            if mode == "ranking":
-                idx = np.argmax(one_hot)  # Get the current one-hot value
-
-                if idx in freeze_indices:  # do not augment
-                    return sample_dict
-
-                # with prob 0.5, do: TODO make it more beautiful :)
-                if random.random() < 0.5:
-                    # raise rank
-                    idx += 1
-                    idx = min(idx, len(one_hot) - 1)
-                else:
-                    # decrease rank
-                    idx -= 1
-                    idx = max(idx, 0)
 
         return sample_dict
